박은빈 2019.12.24 SongVo int -> long 변경

+++ MusiColor/src/main/java/com/musicolor/www/vo/SongVO.java
@@ -4,27 +4,26 @@ import org.springframework.web.multipart.MultipartFile;
 
 public class SongVO { 
 	// song 테이블
-	private int s_no; // 노래번호
+	private long s_no; // 노래번호
 	private int s_vno; // 가수번호
 	private String s_title; // 노래제목
-	private int s_ano; // 사진번호
+	private long s_ano; // 사진번호
 	
 	// vocal 테이블
 	private String v_name; // 가수 이름
 	
 	// youtube 테이블
-	private int y_no; // 유튜브 번호
+	private long y_no; // 유튜브 번호
 	private String y_link; // 유튜브 링크
 	private int y_mno; // 업로드 회원 번호
-	
 	// 파일 업로드 관련
 	private MultipartFile sFile;
 	private MultipartFile[] file;
 	
-	public int getS_no() {
+	public long getS_no() {
 		return s_no;
 	}
-	public void setS_no(int s_no) {
+	public void setS_no(long s_no) {
 		this.s_no = s_no;
 	}
 	public int getS_vno() {
@@ -39,10 +38,10 @@ public class SongVO {
 	public void setS_title(String s_title) {
 		this.s_title = s_title;
 	}
-	public int getS_ano() {
+	public long getS_ano() {
 		return s_ano;
 	}
-	public void setS_ano(int s_ano) {
+	public void setS_ano(long s_ano) {
 		this.s_ano = s_ano;
 	}
 	public String getV_name() {
@@ -51,10 +50,10 @@ public class SongVO {
 	public void setV_name(String v_name) {
 		this.v_name = v_name;
 	}
-	public int getY_no() {
+	public long getY_no() {
 		return y_no;
 	}
-	public void setY_no(int y_no) {
+	public void setY_no(long y_no) {
 		this.y_no = y_no;
 	}
 	public String getY_link() {
